hushh_mcp/agents/Mailer/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
mailjet-rest==1.3.4
